+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 
 Bug fixes:
 
+  - [ct] Fix false diagnostic for missing method #1500
   - [dl] Fix docblock definition location at class level docblocks
   - [idx] Do not try and use non-tokens for property names #1317
   - [cb] Fix rendering of array values in generated code